remove gof.py file and move function into utils.py
