+++ deerlab/utils/__init__.py
@@ -1,4 +1,2 @@
 # __init__.py
 from .utils import *
-from .gof import goodness_of_fit
-
