+++ deerlab/utils/utils.py
@@ -46,15 +46,7 @@ def parse_multidatasets(V_, K, weights, noiselvl, precondition=False, masks=None
     else: 
         subset = subsets.copy()        
 
-    # Noise level estimation/parsing
-    sigmas = np.zeros(len(subset))
-    if noiselvl is None:
-        for i in range(len(subset)):
-            sigmas[i] = der_snr(Vlist[i])
-    else: 
-        noiselvl = np.atleast_1d(noiselvl)
-        sigmas = noiselvl.copy()
-
+    # Parse the masks
     if masks is None: 
         masks = [np.full_like(V,True).astype(bool) for V in Vlist]
     elif not isinstance(masks,list):
@@ -62,6 +54,16 @@ def parse_multidatasets(V_, K, weights, noiselvl, precondition=False, masks=None
     if len(masks)!= len(Vlist): 
         raise SyntaxError('The number of masks does not match the number of signals.')
     mask = np.concatenate(masks, axis=0) 
+
+    # Noise level estimation/parsing
+    sigmas = np.zeros(len(subset))
+    if noiselvl is None:
+        for i in range(len(subset)):
+            sigmas[i] = der_snr(Vlist[i][masks[i]])
+    else: 
+        noiselvl = np.atleast_1d(noiselvl)
+        sigmas = noiselvl.copy()
+
 
     # Concatenate the datasets along the list axis
     V = np.concatenate(Vlist, axis=0)
@@ -162,6 +164,84 @@ def formatted_table(table,align=None):
     table2print += table_lines # Add a line
     return table2print
 #===============================================================================
+
+
+#===============================================================================
+def goodness_of_fit(x,xfit,Ndof,noiselvl):
+    r""" 
+    Goodness of Fit statistics
+    ==========================
+
+    Computes multiple statistical indicators of goodness of fit.
+
+    Usage:
+    ------
+        stats = goodness_of_fit(x,xfit,Ndof)
+
+    Arguments: 
+    ----------
+    x (N-element array)
+        Original data
+    xfit (N-element array)
+        Fit
+    Ndog (scalar, int)
+        Number of degrees of freedom
+    noiselvl (scalar)
+        Standard dexiation of the noise in x.
+
+    Returns:
+    --------
+    stats (dict)
+        Statistical indicators:
+            stats['chi2red'] - Reduced chi-squared
+            stats['rmsd'] - Root mean-squared dexiation
+            stats['R2'] - R-squared test
+            stats['aic'] - Akaike information criterion
+            stats['aicc'] - Corrected Akaike information criterion
+            stats['bic'] - Bayesian information criterion
+
+    """
+    sigma = noiselvl
+    Ndof = np.maximum(Ndof,1)
+    residuals = x - xfit
+    
+    # Special case: no noise, and perfect fit
+    if np.isclose(np.sum(residuals),0):
+        return {'chi2red':1,'R2':1,'rmsd':0,'aic':-np.inf,'aicc':-np.inf,'bic':-np.inf,'autocorr':0}
+
+    # Get number of xariables
+    N = len(x)
+    # Extrapolate number of parameters
+    Q = Ndof - N
+
+    # Reduced Chi-squared test
+    chi2red = 1/Ndof*np.linalg.norm(residuals)**2/sigma**2
+
+    # Autocorrelation based on Durbin–Watson statistic
+    autocorr_DW = abs( 2 - np.sum((residuals[1:-1] - residuals[0:-2])**2)/np.sum(residuals**2) )
+
+    # R-squared test
+    R2 = 1 - np.sum((residuals)**2)/np.sum((xfit-np.mean(xfit))**2)
+
+    # Root-mean square dexiation
+    rmsd = np.sqrt(np.sum((residuals)**2)/N)
+
+    # Log-likelihood
+    loglike = N*np.log(np.sum((residuals)**2))
+
+    # Akaike information criterion
+    aic =  loglike + 2*Q
+
+    # Corrected Akaike information criterion
+    aicc = loglike + 2*Q + 2*Q*(Q+1)/(N-Q-1)
+
+    # Bayesian information criterion
+    bic =  loglike + Q*np.log(N)
+
+    return {'chi2red':chi2red,'R2':R2,'rmsd':rmsd,'aic':aic,'aicc':aicc,'bic':bic,'autocorr':autocorr_DW}
+#===============================================================================
+
+
 
 
 #===============================================================================
